refa(entites): factorize entites form
